Avoid emitting #line directives within #if blocks
If the block is disabled, then the preprocessor will skip over any #line
directives within it.  The line numbers will be relative to the next
preceding #line directive that isn't skipped, and most likely won't be
the right line.

Fix a few errors in the output:
The name prefix changes YY_@_DEBUG_FLAG, not YY_@_DEBUG.
Output YY_@_LSP_NEEDED, not YYLSP_@_NEEDED.
Delete a spurious "if (debugflag)" that can prevent YY_@_LSP_NEEDED from
being output to the header file when -t/--debug is absent from the
command line.