Fix line number replacement
The code searches for "\n#line @" and replaces the '@' with the line
number.  If it finds a partial match (a few chars in the sequence
followed by one that isn't), then restart the search using the first
character that doesn't match instead of skipping it.

Example: If the input is "\n\n#line @", the second '\n' doesn't match
'#' from the pattern.  It used to ignore this '\n' and reset the search
for the next character in the input, '#', which doesn't match the first
'\n' in the pattern.  As a result it would not match this line and
would not replace the '@' with a line number.