Unlink temporary files immediately after creation
Basic idea is that the temporary file will be automatically removed
whenever the file is closed, in case we execute some code path that
exits without cleaning up.  Likely not supported on MSDOS and VMS, which
was the reason for the extra cleanup steps that got rearranged in the
grandparent commit.  These two probably don't support mkstemp() either
and would need a different solution entirely for temporary files.